aider: Update start script in package.json to use nodemon for server auto-restart.
# Aider chat conversation:

ASSISTANT: To use `nodemon` instead of `node` to start the server, we need to update the `start` script in the `package.json` file. Here's the diff for that change:

```diff
--- backend/package.json
+++ backend/package.json
@@ ... @@
   "scripts": {
-    "start": "node server.js",
+    "start": "nodemon server.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "Shihab Dider",
```

This change will make `nodemon` the default command to run the server, which will automatically restart the server when file changes in the directory are detected.